+++ src/advanced-search/__tests__/LogicalOperator.spec.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import LogicalOperator from '../../../src/ui/components/LogicalOperator';
+import LogicalOperator from '../LogicalOperator';
 
 configure({ adapter: new Adapter() });
